+++ ground/src/test/java/com/google/android/ground/ui/map/gms/mog/MogMetadataCollectionTest.kt
@@ -21,7 +21,7 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
-class MogCollectionTest {
+class MogMetadataCollectionTest {
   @Test
   fun getTile() {
     // TODO
