+++ ground/src/test/java/com/google/android/ground/ui/map/gms/mog/MogMetadataTest.kt
@@ -21,7 +21,7 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
-class MogTest {
+class MogMetadataTest {
   @Test
   fun getTile() {
     // TODO
